requirements.txt
aiogram~=3.22.0
python-dotenv~=1.2.1
groq~=0.25.0
pandas~=2.2.0
openpyxl~=3.1.0
